feat: move compilation into test
