+++ crates/blockifier/src/test_utils/struct_impls.rs
@@ -1,5 +1,6 @@
 use std::sync::Arc;
 
+use cairo_native::executor::AotNativeExecutor;
 use cairo_vm::vm::runners::cairo_runner::ExecutionResources;
 use serde_json::Value;
 use starknet_api::block::{BlockNumber, BlockTimestamp};
@@ -190,6 +191,36 @@ impl ContractClassV1 {
 }
 
 impl NativeContractClassV1 {
+    /// Convenience function to construct a NativeContractClassV1 from a raw contract class.
+    /// If control over the compilation is desired use [Self::new] instead.
+    fn try_from_json_string(raw_contract_class: &str) -> Result<Self, Box<dyn std::error::Error>> {
+        // Compile the Sierra Program to native code and loads it into the process'
+        // memory space.
+        fn compile_and_load(
+            sierra_program: &cairo_lang_sierra::program::Program,
+        ) -> Result<AotNativeExecutor, cairo_native::error::Error> {
+            let native_context = cairo_native::context::NativeContext::new();
+            let native_program = native_context.compile(sierra_program, None)?;
+            Ok(AotNativeExecutor::from_native_module(
+                native_program,
+                cairo_native::OptLevel::Default,
+            ))
+        }
+
+        let sierra_contract_class: cairo_lang_starknet_classes::contract_class::ContractClass =
+            serde_json::from_str(raw_contract_class)?;
+
+        // todo(rodro): we are having two instances of a sierra program, one it's object form
+        // and another in its felt encoded form. This can be avoided by either:
+        //   1. Having access to the encoding/decoding functions
+        //   2. Refactoring the code on the Cairo mono-repo
+
+        let sierra_program = sierra_contract_class.extract_sierra_program()?;
+        let executor = compile_and_load(&sierra_program)?;
+
+        Ok(Self::new(&sierra_program, executor, sierra_contract_class)?)
+    }
+
     pub fn from_file(contract_path: &str) -> Self {
         let raw_contract_class = get_raw_contract_class(contract_path);
         Self::try_from_json_string(&raw_contract_class).unwrap()
